Improve query engine benchmark setup
Improve the query engine benchmark setup to allow running benchmarks
with some optimizations toggled on and off. Poll the external view to
wait until segments are loaded instead of having a fixed timeout. Add
support for multiple query strings in the benchmark, doing a cross
product of optimization flags and queries. Add option to output
profiling information. Update the version of JMH.